fixes #22
Implemented LoggingLevelLabel as viewable component in the status bar. A click onto the label will open a popup menu that allows the user to set the logging level for the current process.

The Version number of EvA2 has now increased to 3.0-rc1. Due to the major changes in the GUI and server implementation I think the version number 2.x should be dropped to reflect the size of changes.